(feature) change query code generation
